+++ CHANGES.md
@@ -11,7 +11,10 @@ To be released.
 ### Backward-incompatible API changes
 
  -  Changed `BlockLocator.Create()` to take a single `BlockHash`.  [[#3913]]
- -  Removed `threshold` parameter from `BlockChain.GetBlockLocator()`  [[#3913]]
+ -  Changed `BlockLocator()` to throw an `ArgumentException` when
+    the number of given `BlockHash`es is not 1.  [[#3913]]
+ -  Removed `threshold` parameter from `BlockChain.GetBlockLocator()`.
+    [[#3913]]
  -  (Libplanet.Net) Removed `SwarmOptions.BranchpointThreshold` property.
     [[#3913]]
 
@@ -31,6 +34,8 @@ To be released.
 ### Dependencies
 
 ### CLI tools
+
+[#3913]: https://github.com/planetarium/libplanet/pull/3913
 
 
 Version 5.2.2